+++ coupon-platform-domain/src/main/kotlin/coupon_platform/domain/issued_coupon/service/IssuedCouponStoreService.kt
@@ -1,7 +1,7 @@
 package coupon_platform.domain.issued_coupon.service
 
 import coupon_platform.domain.common.CommonConstants.Companion.EXTERNAL_ID_LENGTH
-import coupon_platform.domain.coupon_code.SuspendableRandomNumberGenerator
+import coupon_platform.domain.common.SuspendableRandomNumberGenerator
 import coupon_platform.domain.issued_coupon.dto.IssueCouponCommand
 import coupon_platform.domain.issued_coupon.repository.IssuedCouponStore
 import org.springframework.beans.factory.annotation.Qualifier
